Create working oauth CLI.
Figured out how to create working oauth. Should implement tomorrow.
Basically a window with a label explaining what to do and a button to
launch the authentication function (and hence the browser). There should
be an entry to copy the link as well. Figure out how to know whether
we've been authenticated. Resulting token and secret should be stored in
a database. Window should also show username and handle of currently
authenticated user. In the far future, It is best to implement a user
selector from the already authenticated users.