Implement database, twitter stream and authentication.
Implemented a basic sqllite database for this system. An extra table
with hashtags should be added to allow for tracking hashtag popularity.
Twitter streaming functionality added.
Ouath now saves the settings resultant from authentication for later use
and sets them as default.
Ouath module should be meshed with twitter module, and both added to the
GUI. This should not be too hard.